+++ iniesta/loggers.py
@@ -1,0 +1,6 @@
+import logging
+
+
+logger = logging.getLogger('root.iniesta')
+error_logger = logging.getLogger('sanic.error.iniesta')
+access_logger = logging.getLogger('sanic.access.iniesta')