+++ test/test_sns.py
@@ -221,35 +221,35 @@ class TestSNSClient(SNSInfra):
     async def test_client_initialize(self, start_local_aws, create_global_sns, sns_endpoint_url):
 
         client = await SNSClient.initialize(
-            topic_arn=create_global_sns['TopicArn'],
-            endpoint_url=sns_endpoint_url)
+            topic_arn=create_global_sns['TopicArn'])
         assert client.topic_arn == create_global_sns['TopicArn']
 
 
-    async def test_client_publish(self, start_local_aws, create_global_sns, sns_endpoint_url):
+    async def test_client_create_message(self, start_local_aws, create_global_sns, sns_endpoint_url):
         client = await SNSClient.initialize(
-            topic_arn=create_global_sns['TopicArn'],
-            endpoint_url=sns_endpoint_url
+            topic_arn=create_global_sns['TopicArn']
         )
 
-        response = await client.publish_event(event="SomethingAwesomeHappened", message="Great Success!", value="Something")
+        response = client.create_message(
+            event="SomethingAwesomeHappened",
+            message="Great Success!", value="Something")
 
         assert response is not None
-        assert 'MessageId' in response
-        assert response['MessageId'] is not None
+        assert response.client == client
+        assert isinstance(response, SNSMessage)
+        # assert 'MessageId' in response
+        # assert response['MessageId'] is not None
 
     async def test_client_topic_doesnt_exist(self, start_local_aws, sns_endpoint_url):
 
         with pytest.raises(botocore.exceptions.ClientError):
             client = await SNSClient.initialize(
-                topic_arn="asdasda",
-                endpoint_url=sns_endpoint_url)
+                topic_arn="asdasda")
 
     async def test_list_subscriptions_empty(self, start_local_aws, create_global_sns, sns_endpoint_url):
 
         client = await SNSClient.initialize(
-            topic_arn=create_global_sns['TopicArn'],
-            endpoint_url=sns_endpoint_url,
+            topic_arn=create_global_sns['TopicArn']
         )
 
         subscriptions = [s async for s in client.list_subscriptions_by_topic()]
@@ -261,8 +261,7 @@ class TestSNSClient(SNSInfra):
                                                     sns_endpoint_url):
 
         client = await SNSClient.initialize(
-            topic_arn=create_global_sns['TopicArn'],
-            endpoint_url=sns_endpoint_url
+            topic_arn=create_global_sns['TopicArn']
         )
 
         subscriptions = [s async for s in client.list_subscriptions_by_topic()]
@@ -298,8 +297,7 @@ class TestSNSClient(SNSInfra):
         """
 
         client = await SNSClient.initialize(
-            topic_arn=create_global_sns['TopicArn'],
-            endpoint_url=sns_endpoint_url
+            topic_arn=create_global_sns['TopicArn']
         )
 
         subscriptions = [s async for s in client.list_subscriptions_by_topic()]